requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
mysql-connector-python>=8.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
